dom/animation/test/css-transitions/test_effect-name.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var div = addDiv(t);

  // Add a transition
  div.style.left = '0px';
  window.getComputedStyle(div).transitionProperty;
  div.style.transition = 'all 100s';
  div.style.left = '100px';

  assert_equals(div.getAnimations()[0].effect.name, 'left',
                'The name for the transitions corresponds to the property ' +
                'being transitioned');
}, 'Effect name for transitions');

</script>
